+++ tortoise/backends/base_postgres/schema_generator.py
@@ -62,7 +62,7 @@ class BasePostgresSchemaGenerator(BaseSchemaGenerator):
         auto_now: bool = False,
     ) -> str:
         default_str = " DEFAULT"
-        default_str += " CURRENT_TIMESTAMP" if auto_now_add else f" {default}"
+        default_str += " CURRENT_TIMESTAMP" if (auto_now_add or auto_now) else f" {default}"
         return default_str
 
     def _escape_default_value(self, default: Any):