+++ tortoise/backends/mysql/schema_generator.py
@@ -62,7 +62,7 @@ class MySQLSchemaGenerator(BaseSchemaGenerator):
         default_str = " DEFAULT"
         if not (auto_now or auto_now_add):
             default_str += f" {default}"
-        if auto_now_add:
+        if auto_now_add or auto_now:
             default_str += " CURRENT_TIMESTAMP(6)"
         if auto_now:
             default_str += " ON UPDATE CURRENT_TIMESTAMP(6)"