+++ tortoise/backends/sqlite/schema_generator.py
@@ -32,7 +32,7 @@ class SqliteSchemaGenerator(BaseSchemaGenerator):
         auto_now: bool = False,
     ) -> str:
         default_str = " DEFAULT"
-        default_str += " CURRENT_TIMESTAMP" if auto_now_add else f" {default}"
+        default_str += " CURRENT_TIMESTAMP" if (auto_now_add or auto_now) else f" {default}"
         return default_str
 
     def _escape_default_value(self, default: Any):